requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-jose
passlib==1.7.4
bcrypt==3.2.2
python-multipart
pytest
pytest-asyncio
httpx
email-validator
dotenv